Refactor: Clean up test files and update configurations
- Remove Playwright e2e tests and related files
- Clean up test helpers and production test scripts
- Update ESLint configuration
- Add unit tests for worker services
- Update package.json dependencies
- Remove unused SQL files and test logs
