+++ package.json
@@ -12,15 +12,10 @@
 		"lint:fix": "eslint . --fix",
 		"lint:check": "eslint . --max-warnings 0",
 		"test": "vitest run",
-		"test:all": "npm run test && npm run test:e2e",
+		"test:all": "npm run test",
 		"test:watch": "vitest --watch",
 		"test:ui": "vitest --ui",
-		"test:coverage": "vitest --coverage",
-		"test:e2e": "playwright test",
-		"test:e2e:ui": "playwright test --ui",
-		"test:e2e:headed": "playwright test --headed",
-		"test:e2e:debug": "playwright test --debug",
-		"test:e2e:report": "playwright show-report"
+		"test:coverage": "vitest --coverage"
 	},
 	"dependencies": {
 		"@heroicons/react": "^2.2.0",
@@ -44,7 +39,6 @@
 		"@cloudflare/workers-types": "^4.20250715.0",
 		"@eslint/compat": "^1.4.0",
 		"@eslint/js": "^9.36.0",
-		"@playwright/test": "^1.55.0",
 		"@preact/preset-vite": "^2.9.3",
 		"@testing-library/jest-dom": "^6.6.3",
 		"@testing-library/preact": "^3.2.4",
