--- sync_teams.sql
@@ -1,11 +1,0 @@
-UPDATE teams SET 
-  slug = 'north-carolina-legal-services',
-  name = 'North Carolina Legal Services',
-  config = '{"aiModel":"llama","consultationFee":75,"requiresPayment":true,"ownerEmail":"paulchrisluke@gmail.com","availableServices":["Family Law","Small Business and Nonprofits","Employment Law","Tenant Rights Law","Probate and Estate Planning","Special Education and IEP Advocacy"],"domain":"northcarolinalegalservices.blawby.com","description":"Affordable, comprehensive legal services for North Carolina. Family Law, Small Business, Employment, Tenant Rights, Probate, Special Education, and more.","paymentLink":"https://app.blawby.com/northcarolinalegalservices/pay?amount=7500","brandColor":"#059669","accentColor":"#10b981","introMessage":"Welcome to North Carolina Legal Services! I''m here to help you with affordable legal assistance in areas including Family Law, Small Business, Employment, Tenant Rights, Probate, and Special Education. I can answer your questions and help you connect with our experienced attorneys. How can I assist you today?","profileImage":"https://app.blawby.com/storage/team-photos/uCVk3tFuy4aTdR4ad18ibmUn4nOiVY8q4WBgYk1j.jpg"}'
-WHERE id = '01jq70jnstyfzevc6423czh50e';
-
-UPDATE teams SET 
-  slug = 'blawby-ai',
-  name = 'Blawby AI',
-  config = '{"aiModel":"llama","consultationFee":0,"requiresPayment":false,"ownerEmail":"paulchrisluke@gmail.com","availableServices":["Case Discussion & Analysis","Document Organization","Timeline & Evidence Gathering","Case Summary Preparation","Attorney-Ready Documentation"],"serviceQuestions":{"Case Discussion & Analysis":["What legal issue are you dealing with?","When did this situation first begin?","What''s the current status of your case?","Have you taken any legal action yet?"],"Document Organization":["What documents do you have related to this case?","Do you have any contracts, emails, or written communications?","Are there any photos, videos, or other evidence?","What''s the most important document you have?"],"Timeline & Evidence Gathering":["Can you walk me through what happened chronologically?","Who else was involved in this situation?","What evidence do you have to support your case?","Are there any witnesses who can support your version of events?"],"Case Summary Preparation":["What outcome are you hoping to achieve?","What''s your ideal timeline for resolving this?","Have you consulted with any attorneys before?","What questions do you have for an attorney?"],"Attorney-Ready Documentation":["What type of legal matter is this?","Do you have all the necessary documents organized?","What specific help do you need preparing for your attorney meeting?","Are there any deadlines or time-sensitive aspects to your case?"]},"domain":"ai.blawby.com","description":"AI-powered legal case preparation assistant. Discuss your legal situation, organize your case details, and prepare a comprehensive case summary ready for attorney review.","paymentLink":null,"brandColor":"#2563eb","accentColor":"#3b82f6","introMessage":"Hello! I''m Blawby AI, your legal case preparation assistant. I''m here to help you organize your legal situation, gather all the important details, and prepare a comprehensive case summary that you can take to any attorney. Think of me as your case preparation partner - I''ll help you get your story straight and make sure nothing important gets missed. What legal issue would you like to discuss today?","profileImage":null,"blawbyApi":{"enabled":false,"apiUrl":"https://staging.blawby.com"}}'
-WHERE id = '01K0TNGNKTM4Q0AG0XF0A8ST0Q';
