+++ wrangler.toml
@@ -77,6 +77,6 @@ d1_databases = [
 [[env.production.queues.producers]]
 queue = "doc-events"
 binding = "DOC_EVENTS"
-# r2_buckets = [
-#   { binding = "FILES_BUCKET", bucket_name = "blawby-ai-files" }
-# ] 
+[[env.production.r2_buckets]]
+binding = "FILES_BUCKET"
+bucket_name = "blawby-ai-files"
